Refactor oper application expression in ToIrListener
Towards #54

In order to make the representation of identifiers abstract, we're gonna
end up parsing the `opcode`s here into the `Identifier` structure.
DRYing out the construction of the applications will make this change
much less invasive. To that end, this PR just factors out the most oft
repeated construction of the operator applications.