Add error handling for file loading
Closes #215

This is introduced in a way to suggest turning to sweet-monads/either
for sequencing our error-prone computations.

The sequencing and composition of our top-level routines exposed by CLI
is currently done via imperative early exits. This makes it hard to
reason about the code in general, but also breaks our ability to compose
these routines. Looking towards #345, we'll need typecheck to receive
the output of parse, then add it's typing information, before finally
printing to the `--out` location. This requires composing the routines
while handling possible error states, and just bailing with exit codes
won't work for that. Since we are already using `Either`, and since it
is designed for facilitating the composition of error-prone code, I
think we can massage these routines into shape with a few well-placed
Either's. ;)