Introduce an ItfTrace type
For #888, we'll need to support converting ITF into `QuintEx`. Having a
type to specify the right shape for the ITF will make the conversion a
nobrainer, and help ease any other manipulation of ITF data in the
future.

This also adds some minor cleanup to the ITF representation to make it
more faithful: namely, number are numbers, and bitings are encoded as
strings.