Refactor ToIrListener id generation
Add a helper function to generate location IDs and register them into
the source map. These must always go together, so adding the two into
the same function saves a number of lines and makes the code safer since
it ensures we won't miss a needed addition to the map.