Rework wilcard case elim to use unary lambda
We had been using thunks, lambdas that take no arguments, to encode the
eliminators for wildcard match cases. I.e., the match expression

```
match foo {
  | _ => bar
}
```

Was being converted into our builtin exotic thus:

```
matchVariant(foo, "_", () => bar)
```

This made sense because in the case of the wildcard match we are
discarding the variant, and there is no argument to be supplied to the
eliminator.

However:

1. We don't really support nullary lambdas (even tho we are making use
of them at several places in our code currently).
2. Since we are committed to using the exotic builtin operators, and there
is no way to specify the effect of a variadic builtin with unpredictable
arguments, we could not give this builtin the correct effect signature.

So I have changed the way wildcard cases are encoded, to also use unary
lambas that just ignore their argument. This simplifies a lot of the
match typing and evaluation, so it is the right call overall.