Sort constraints to solve by order of "generality"
Closes #1275

Since we don't do any backtracking in our constraint solving, the order
in which the are solved is important. In particular, if we solve
`isDefined` constraints before solving equality constraints, we can end
up unifying a free variable with a variable in the first discovered
defined type, rather than finding the most specific type that can unify
with the given schema.

This change defines an order over constraints, based on their kind, and
uses that to sort the constraints before they are solved.