Ensure transformations can't mutate the original object
There was some mutation happening via the transformer, which was
wreaking havoc on attempts to complete #1298.

Apparently it was also somehow leading to incorrect printing of sum
type variants. I haven't had time to investigate why, but deep cloning
the original object must be safer and it is fixing a blocking problem in
ongoing work in #1298.

Since the cases in which this causes problems are quite subtle, I havent
been able to figure out how to test this directly.