Clarify `quint test` usage in the Alpenglow post
The current wording leads inexperienced users to believe that by running
`quint test statemachine.qnt` all tests in the file will be executed,
which isn't true. The test command must be called with the spec file and
the main module in this example.

Moreover, it mentions that `quint test` can be used in CI to run tests,
which is true but, the wording is overloaded by other commonly used
tooling, like `npm test` for example, where the expectation is that by
simply running `test` it runs all tests in the codebase. That isn't true
for quint (yet), so it's worth expanding to the full command one would
use in CI right now.