Fix copy and past for multi-line expression
The repl supports trimming the prompt and the continue prompt prefixes
when copying and pasting previous expressions form the repl. However,
the '...' prefix of multi-line expressions confuses the interpreter into
thinking that those lines are special commands instead.

This patches fixes the repl's line handler so that lines starting with
'...' continue prompt prefix are interpreted as quint expressions, thus
allowing for copying and pasting multi-line commands from the repl
itself.