Introduce support for repl eval expressions
Currently, the CLI does not support opening the repl with a expression
to evaluate. Users managed to work around that with linux pipes where
they first `echo` the command they want to run, then pipe the result to
the quint process. Although this approach works, it's less ergonomic and
causes LLMs to ask for confirmation before running commands in the repl
due to mistrust in the `echo` command.

This patch introduces support to calling quint repl with a command to be
evaluated upon entering the repl.